requirements.txt
pandas==1.3.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
numpy==1.24.3
joblib==1.1.1
scikit-learn==1.2
seaborn==0.11.1
xgboost==1.4.2
imblearn==0.0
missingno==0.5.2
catboost==1.2
propy3==1.1.1
